refactor file structure and naming
