+++ backend/src/modules/common/base-entity/base-entity.class.ts
