+++ backend/src/modules/common/index.ts
@@ -1,3 +1,4 @@
+export * from './base-entity/base-entity.class';
 export * from './controller/controller.factory';
 export * from './repository/repository.factory';
 export * from './service/service.factory';
