+++ backend/src/modules/experience/router.ts
@@ -2,7 +2,7 @@ import express from 'express';
 import type { EntitySchema } from 'typeorm';
 import dataSource from '../../dataSource';
 import { createController, createRepository, createService } from '../common';
-import { Experience } from './entity';
+import { Experience } from './entity.class';
 
 const dataRepository = createRepository<Experience>(
   Experience as unknown as EntitySchema<Experience>,
