+++ backend/src/modules/ministry/entity.class.ts
@@ -1,5 +1,5 @@
 import { Column, Entity } from 'typeorm';
-import { BaseEntity } from '../common/base-entity';
+import { BaseEntity } from '../common';
 
 @Entity()
 export class Ministry extends BaseEntity {
