+++ backend/src/modules/user/entity.class.ts
@@ -1,5 +1,5 @@
 import { Entity, Column } from 'typeorm';
-import { BaseEntity } from '../common/base-entity';
+import { BaseEntity } from '../common';
 
 @Entity()
 export class User extends BaseEntity {
