+++ backend/src/modules/user/router.ts
@@ -1,7 +1,7 @@
 import express from 'express';
 import dataSource from '../../dataSource';
 import { createController, createRepository, createService } from '../common';
-import { User } from './entity';
+import { User } from './entity.class';
 
 import type { EntitySchema } from 'typeorm';
